added Particle.Radius; deteled Particle.Radius(); changed projectile to have Radius of 0.
Having the Projectile have a radius of 0 is almost like having a camera with a tiny pin-hole aperture (high F-number). When the projectile's radius is set to smaller numbers, the image becomes sharper. This is because the smaller projectile can move around the map more because it is less likely to collide with stationary targets.